refactor worker scheduling
